requirements.txt
certifi==2023.5.7
charset-normalizer==3.2.0
contourpy==1.1.0
cycler==0.11.0
EasyProcess==1.1
entrypoint2==1.1
filelock==3.12.2
fonttools==4.41.0
idna==3.4
importlib-resources==6.0.0
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
MouseInfo==0.1.3
mpmath==1.3.0
mss==9.0.1
networkx==3.1
numpy==1.25.1
opencv-python==4.8.0.74
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
psutil==5.9.5
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pynput==1.7.6
pyobjc-core==9.2
pyobjc-framework-ApplicationServices==9.2
pyobjc-framework-Cocoa==9.2
pyobjc-framework-Quartz==9.2
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
python-dateutil==2.8.2
pytweening==1.0.7
pytz==2023.3
PyYAML==6.0
requests==2.31.0
rubicon-objc==0.4.6
scipy==1.11.1
seaborn==0.12.2
six==1.16.0
sympy==1.12
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
typing_extensions==4.7.1
tzdata==2023.3
ultralytics==8.0.133
urllib3==2.0.3
zipp==3.16.0
